+++ htdocs/database/location.php
@@ -5,7 +5,6 @@ require_once "../modules/.php";
 // IMPORTS
 module("functional", "check");
 module("functional", "database");
-module("functional", "environment");
 module("functional", "post");
 
 // SIGNAL
@@ -15,26 +14,47 @@ signal("functional");
 header('Content-Type: application/json');
 header('Access-Control-Allow-Origin: *');
 
-// CONNECT TO THE DATABASE
-$database = database_connect($ENV["DB_HOST"], $ENV["DB_USER"], $ENV["DB_PASSWORD"], $ENV["DB_NAME"]);
-
 // CHECK ARGUMENTS
-check('/^[a-z]{2}$/', $PST["LANG"], "LANG");
-check('/^[A-Z0-9]{6}$/', $PST['ID'], "ID");
-check('/^[A-Z0-9]{4}$/', $PST["NODE"], "NODE");
-check('/^[A-Z0-9]{2}$/', $PST["CLUSTER"], "CLUSTER");
+check($PST, "LANG");
+check($PST, "ID");
+check($PST, "NODE");
+check($PST, "CLUSTER");
 
 // CHECK ARGUMENT RELATIONSHIPS
-if (!(array_key_exists("LANG", $PST) and (array_key_exists("ID", $PST) xor array_key_exists("NODE", $PST) xor array_key_exists("CLUSTER", $PST)))) {
+if (!(isx($PST, "LANG") and (isx($PST, "ID") xor isx($PST, "NODE") xor isx($PST, "CLUSTER")))) {
     throw new incorrectArgumentsError();
 }
 
 // TYPES OF QUERIES
-if (array_key_exists("ID", $PST)) {
-    $problemsQuery = database_request("SELECT node, name_" . $PST["LANG"] . " FROM problems WHERE id = '" . $PST["ID"] . "'", $database)->fetch_assoc();
-    $nodesQuery = database_request("SELECT cluster, name_" . $PST["LANG"] . " FROM nodes WHERE node = '" . $problemsQuery["node"] . "'", $database)->fetch_assoc();
-    $clustersQuery = database_request("SELECT tree, name_" . $PST["LANG"] . " FROM clusters WHERE cluster = '" . $nodesQuery["cluster"] . "'", $database)->fetch_assoc();
-    $treesQuery = database_request("SELECT name_" . $PST["LANG"] . " FROM trees WHERE tree = '" . $clustersQuery["tree"] . "'", $database)->fetch_assoc();
+if (isx($PST, "ID")) {
+    $problemsQuery = database_request(
+        "SELECT node, ? FROM problems WHERE id = ?",
+        [
+            "name_" . $PST["LANG"],
+            $PST["ID"]
+        ]
+    )->fetch_assoc();
+    $nodesQuery = database_request(
+        "SELECT cluster, ? FROM nodes WHERE node = ?",
+        [
+            "name_" . $PST["LANG"],
+            $problemsQuery["node"]
+        ]
+    )->fetch_assoc();
+    $clustersQuery = database_request(
+        "SELECT tree, ? FROM clusters WHERE cluster = ?",
+        [
+            "name_" . $PST["LANG"],
+            $nodesQuery["cluster"]
+        ]
+    )->fetch_assoc();
+    $treesQuery = database_request(
+        "SELECT ? FROM trees WHERE tree = ?",
+        [
+            "name_" . $PST["LANG"],
+            $clustersQuery["tree"]
+        ]
+    )->fetch_assoc();
     echo json_encode([
         "id" => [
             "value" => $PST["ID"],
@@ -53,10 +73,29 @@ if (array_key_exists("ID", $PST)) {
             "name" => $treesQuery["name_" . $PST["LANG"]]
         ]
     ]);
-} elseif (array_key_exists("NODE", $PST)) {
-    $nodesQuery = database_request("SELECT cluster, name_" . $PST["LANG"] . " FROM nodes WHERE node = '" . $PST["NODE"] . "'", $database)->fetch_assoc();
-    $clustersQuery = database_request("SELECT tree, name_" . $PST["LANG"] . " FROM clusters WHERE cluster = '" . $nodesQuery["cluster"] . "'", $database)->fetch_assoc();
-    $treesQuery = database_request("SELECT name_" . $PST["LANG"] . " FROM trees WHERE tree = '" . $clustersQuery["tree"] . "'", $database)->fetch_assoc();
+} elseif (isx($PST, "NODE")) {
+    $nodesQuery = database_request(
+        "SELECT cluster, ? FROM nodes WHERE node = ?",
+        [
+            "name_" . $PST["LANG"],
+            $PST["NODE"]
+        ]
+    )->fetch_assoc();
+    $clustersQuery = database_request(
+        "SELECT tree, name_" . $PST["LANG"] . " FROM clusters WHERE cluster = '" . $nodesQuery["cluster"] . "'",
+        "SELECT tree, ? FROM clusters WHERE cluster = ?",
+        [
+            "name_" . $PST["LANG"],
+            $nodesQuery["cluster"]
+        ]
+    )->fetch_assoc();
+    $treesQuery = database_request(
+        "SELECT ? FROM trees WHERE tree = ?",
+        [
+            "name_" . $PST["LANG"],
+            $clustersQuery["tree"]
+        ]
+    )->fetch_assoc();
     echo json_encode([
         "id" => [
             "value" => null,
@@ -76,8 +115,20 @@ if (array_key_exists("ID", $PST)) {
         ]
     ]);
 } else {
-    $clustersQuery = database_request("SELECT tree, name_" . $PST["LANG"] . " FROM clusters WHERE cluster = '" . $PST["CLUSTER"] . "'", $database)->fetch_assoc();
-    $treesQuery = database_request("SELECT name_" . $PST["LANG"] . " FROM trees WHERE tree = '" . $clustersQuery["tree"] . "'", $database)->fetch_assoc();
+    $clustersQuery = database_request(
+        "SELECT tree, ? FROM clusters WHERE cluster = ?",
+        [
+            "name_" . $PST["LANG"],
+            $PST["CLUSTER"]
+        ]
+    )->fetch_assoc();
+    $treesQuery = database_request(
+        "SELECT ? FROM trees WHERE tree = ?",
+        [
+            "name_" . $PST["LANG"],
+            $clustersQuery["tree"]
+        ]
+    )->fetch_assoc();
     echo json_encode([
         "id" => [
             "value" => null,