+++ htdocs/modules/f-cryptography.php
@@ -1,12 +1,16 @@
 <?php
-// ENCRYPT FUNCTION
+/*                                                                           */
+/* FUNCTIONS                                                                 */
+/*                                                                           */
+
+// FUNCTIONS -> ENCRYPT
 function encrypt(string $data, string $key, string $method = 'aes-256-cbc'): string {
     $vector = openssl_random_pseudo_bytes(openssl_cipher_iv_length($method));
     $cipher = openssl_encrypt($data, $method, $key, OPENSSL_RAW_DATA, $vector);
     return base64_encode($vector . $cipher);
 }
 
-// DECRYPT FUNCTION
+// FUNCTIONS -> DECRYPT
 function decrypt(string $cipher, string $key, string $method = 'aes-256-cbc'): string {
     $raw = base64_decode($cipher);
     $vector = openssl_cipher_iv_length($method);