+++ htdocs/modules/f-post.php
@@ -1,5 +1,9 @@
 <?php
-// GET POST DATA
+/*                                                                           */
+/* POST                                                                      */
+/*                                                                           */
+
+// POST -> FUNCTION
 global $PST;
 foreach (json_decode(file_get_contents("php://input")) as $key => $value) {
     $PST[strtoupper($key)] = $value;